remove some functions from MapActivity to declutter
